refactor attachment code into its own file
