+++ lib/inventree/stock.dart
@@ -3,6 +3,7 @@ import "dart:async";
 import "package:flutter/material.dart";
 import "package:inventree/api.dart";
 import "package:inventree/helpers.dart";
+import "package:inventree/inventree/attachment.dart";
 import "package:inventree/l10.dart";
 
 import "package:inventree/inventree/part.dart";
