+++ lib/widget/attachment_widget.dart
@@ -2,6 +2,7 @@ import "dart:io";
 
 import "package:flutter/material.dart";
 import "package:flutter_tabler_icons/flutter_tabler_icons.dart";
+import "package:inventree/inventree/attachment.dart";
 import "package:one_context/one_context.dart";
 
 import "package:inventree/api.dart";
